Hashing: use a 64-bit storage type on all platforms.
`size_t` varying across platforms can cause issues, for example mismatching
hashes when cross-compiling modules from a 64bit target to a 32bit target.

Although the comments note that `hash_code` values "are not stable to save or
persist", it is effectively used in such cases today, for example for implicit
module hashes.

Since hashing mechanisms in practice already use `uint64_t` for computations,
use `uint64_t` instead of `size_t` to store the value of `hash_code`.
This is similar to earlier changes in c0445098519defc4bd13624095fac92977c9cfe4.

+++ clang/include/clang/AST/DataCollection.h
@@ -51,7 +51,8 @@ template <class T> void addDataToConsumer(T &DataConsumer, const QualType &QT) {
 
 template <class T, class Type>
 std::enable_if_t<std::is_integral<Type>::value || std::is_enum<Type>::value ||
-                 std::is_convertible<Type, size_t>::value // for llvm::hash_code
+                 std::is_convertible<Type, uint64_t>::value // for
+                                                            // llvm::hash_code
                  >
 addDataToConsumer(T &DataConsumer, Type Data) {
   DataConsumer.update(StringRef(reinterpret_cast<char *>(&Data), sizeof(Data)));
